Authorization added

+++ backend/routes/posts.js
@@ -35,7 +35,8 @@ router.post("",checkAuth, multer({storage: storage}).single("image"), (req, res,
     const post = new Post({
         title: req.body.title,
         content: req.body.content,
-        imagePath: url + "/images/" + req.file.filename
+        imagePath: url + "/images/" + req.file.filename,
+        creator: req.userData.userId
     });
     post.save().then(createdPost => {
         //console.log(result);
@@ -61,13 +62,18 @@ router.put("/:id",checkAuth, multer({storage: storage}).single("image"), (req, r
         _id: req.body.id,
         title: req.body.title,
         connect: req.body.content,
-        imagePath: imagePath
+        imagePath: imagePath,
+        creator: req.userData.userId
     });
     console.log(post);
-    Post.updateOne({_id: req.params.id},post)
+    Post.updateOne({_id: req.params.id, creator: req.userData.userId},post)
     .then(result => {
         console.log(result);
-        res.status(200).json({message: 'Update Successful'});
+        if(result.nModified > 0){
+            res.status(200).json({message: 'Update Successful'});
+        }else {
+            res.status(401).json({message: 'User not authorized'});
+        }
     });
 });
 
@@ -111,10 +117,14 @@ router.get("/:id", (req, res, next) => {
 
 router.delete("/:id",checkAuth, (req, res, next) =>{
     console.log(req.params.id);
-    Post.deleteOne({_id: req.params.id})
+    Post.deleteOne({_id: req.params.id, creator: req.userData.userId})
     .then( result => {
         console.log(result);
-        res.status(200).json({message: 'Post Deleted'});
+        if(result.n > 0){
+            res.status(200).json({message: 'Post Deleted Successfully'});
+        }else {
+            res.status(401).json({message: 'User not authorized'});
+        }
     });
 });
